requirements.txt
Django
psycopg2
django-markdown-app
django-grappelli
django-tabbed-admin
django-countries-plus
django-apptemplates
